Terminal improvements (automagik-forge 29d6c7f1)
Rework the #terminal-container component entirely, making it more responsive and similar to a actual ubuntu terminal.

Use the attached images as reference for building and referencing components.
 ![image.png](/api/images/9755b12c-ca93-4cb9-9f41-16555fdb4368/file)
 ![image.png](/api/images/48fb9eba-f104-4d35-92b3-be4a04eab1d0/file)
 ![image.png](/api/images/64a962f4-3cb1-4f4e-beac-8e155894c122/file)

Explore any shadc-UI components that could help the new implementation.
Try, for instance experiment with the Menubar (https://www.shadcn-vue.com/docs/components/menubar) component for the menu items on the terminal.

As a guideline, make the terminal occupy about 40% of the vertical spacing and about the same horizontal space as currently.

Also try to represente the project "tree structure" in a nicer fashion (like in the images given)